refactor: 타입 구조 및 함수 선언식 공통화
