+++ src/utils/auth.ts
@@ -1,4 +1,4 @@
-export function generateBasicToken(userId: string, userPassword: string): string {
+export const generateBasicToken = (userId: string, userPassword: string): string => {
   const token = btoa(`${userId}:${userPassword}`);
   return `Basic ${token}`;
-}
+};
